requirements.txt
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
pylint==3.0.2
python-json-logger==2.0.7
aiohttp==3.9.1
freezegun==1.4.0
pytest-asyncio==0.23.3 